MK id5 2.4.3

+++ IOTV/src/MK/FreerRTOS/id_2/main/Protocols/embedded/read_write.h
@@ -10,8 +10,7 @@ extern "C"
 
 struct Read_Write
 {
-    //!!! в документации 4 байта!
-    uint64_t dataSize;
+    uint32_t dataSize;
 
     uint8_t nameSize;
     uint8_t channelNumber;
